+++ example_site/app/adapters.py
@@ -8,6 +8,7 @@ class ExampleWorkspaceAdapter(BaseWorkspaceAdapter):
 
     name = "Example workspace"
     type = "example"
+    description = "Example workspace type for demo app"
     list_table_class = tables.ExampleWorkspaceDataTable
     workspace_data_model = models.ExampleWorkspaceData
     workspace_data_form_class = forms.ExampleWorkspaceDataForm